feat: refactor zk-mpc-node models and improve circuit handling; add CLI and proof structures
