+++ packages/zk-mpc-node/Cargo.toml
@@ -4,15 +4,15 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-zk-mpc = { path = "../zk-mpc" }
-ark-bls12-377 = { path = "../zk-mpc/arkworks/curves/bls12_377", version = "0.3.0", features = ["r1cs", "curve"] }
-ark-std = { path = "../zk-mpc/arkworks/std", version = "0.3.0", features = ["std", "print-trace"] }
-ark-marlin = { path = "../zk-mpc/arkworks/marlin", version = "0.3.0", default-features = false, features = [ "std" ] }
-ark-serialize = { path = "../zk-mpc/arkworks/algebra/serialize", version = "0.3.0", default-features = false }
-ark-poly-commit = { path = "../zk-mpc/arkworks/poly-commit", version = "0.3.0" }
-ark-relations = { path = "../zk-mpc/arkworks/snark/relations", version = "0.3.0" }
-mpc-algebra = { path = "../zk-mpc/mpc-algebra", version = "0.1.0" }
-mpc-net = { path = "../zk-mpc/mpc-net", version = "0.1.0" }
+zk-mpc = { path = "../../../zk-mpc" }
+ark-bls12-377 = { path = "../../../zk-mpc/arkworks/curves/bls12_377", version = "0.3.0", features = ["r1cs", "curve"] }
+ark-std = { path = "../../../zk-mpc/arkworks/std", version = "0.3.0", features = ["std", "print-trace"] }
+ark-marlin = { path = "../../../zk-mpc/arkworks/marlin", version = "0.3.0", default-features = false, features = [ "std" ] }
+ark-serialize = { path = "../../../zk-mpc/arkworks/algebra/serialize", version = "0.3.0", default-features = false }
+ark-poly-commit = { path = "../../../zk-mpc/arkworks/poly-commit", version = "0.3.0" }
+ark-relations = { path = "../../../zk-mpc/arkworks/snark/relations", version = "0.3.0" }
+mpc-algebra = { path = "../../../zk-mpc/mpc-algebra", version = "0.1.0" }
+mpc-net = { path = "../../../zk-mpc/mpc-net", version = "0.1.0" }
 
 structopt = "0.3"
 serde = { version = "1.0", features = ["derive"] }
