+++ packages/zk-mpc-node/src/main.rs
@@ -9,13 +9,13 @@ use zk_mpc_node::{models::Opt, node::Node, proof::ProofManager, server::handle_c
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let opt = Opt::from_args();
 
-    // ProofManagerの初期化
+    // Initialize ProofManager
     let proof_manager = Arc::new(ProofManager::new());
 
-    // クライアント接続用のリスナーを作成
+    // Create a listener for client connections
     let listener = TcpListener::bind(format!("127.0.0.1:{}", 9000 + opt.id)).await?;
 
-    // MPCネットワークの初期化
+    // Initialize the MPC network
     let mut net = MPCNetConnection::init_from_path(&opt.input, opt.id as u32);
     net.listen().await.expect("Failed to listen");
     net.connect_to_all()
@@ -24,10 +24,10 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     println!("Listening on port {}", 9000 + opt.id);
 
-    // ノードの初期化
+    // Initialize the node
     let node = Arc::new(Node::new(opt.id as u32, Arc::new(net), proof_manager.clone()).await);
 
-    // クライアントからのリクエストを受け付けるループ
+    // Loop to accept requests from clients
     while let Ok((socket, _)) = listener.accept().await {
         let pm = proof_manager.clone();
         let node = node.clone();
